crud comment-service

+++ video-service/src/main/java/com/hoaithi/video_service/service/VideoService.java
@@ -1,6 +1,7 @@
 package com.hoaithi.video_service.service;
 
 import com.hoaithi.video_service.dto.request.VideoCreationRequest;
+import com.hoaithi.video_service.dto.request.VideoUpdationRequest;
 import com.hoaithi.video_service.dto.response.VideoResponse;
 import com.hoaithi.video_service.entity.Video;
 import com.hoaithi.video_service.entity.VideoHistory;
@@ -8,6 +9,8 @@ import com.hoaithi.video_service.exception.AppException;
 import com.hoaithi.video_service.exception.ErrorCode;
 import com.hoaithi.video_service.mapper.VideoMapper;
 import com.hoaithi.video_service.repository.HistoryRepository;
+import com.hoaithi.video_service.repository.VideoHeartRepository;
+import com.hoaithi.video_service.repository.httpclient.CommentClient;
 import com.hoaithi.video_service.repository.httpclient.FileClient;
 import com.hoaithi.video_service.repository.VideoRepository;
 import lombok.RequiredArgsConstructor;
@@ -32,7 +35,9 @@ public class VideoService {
     VideoRepository videoRepository;
     VideoMapper videoMapper;
     FileClient fileClient;
+    CommentClient commentClient;
     HistoryRepository historyRepository;
+    VideoHeartRepository videoHeartRepository;
 
     public List<VideoResponse> getVideos() {
         List<Video> videos = videoRepository.findAll();
@@ -45,37 +50,43 @@ public class VideoService {
         return videoResponses;
     }
 
-    public VideoResponse getVideoById(String id) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        Video video = videoRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.VIDEO_NOT_EXISTED));
+    public VideoResponse getVideoById(String videoId) {
+        String currentProfileId = getCurrentUserId();
+        Video video = videoRepository.findById(videoId).orElseThrow(() -> new AppException(ErrorCode.VIDEO_NOT_EXISTED));
         video.setViewCount(video.getViewCount()+1);
+        video.setCommentCount(commentClient.countComment(videoId).getResult().getCommentCount());
         historyRepository.save(VideoHistory.builder()
                 .video(video)
-                .userId(authentication.getName())
+                .profileId(currentProfileId)
                 .build());
-        return videoMapper.toVideoResponse(videoRepository.save(video));
+        VideoResponse response = videoMapper.toVideoResponse(videoRepository.save(video));
+        response.setHearted(videoHeartRepository.existsByProfileIdAndVideoId(currentProfileId, videoId));
+        return response;
     }
 
 
-    public Video updateVideo(String id, Video newVideo) {
-        return videoRepository.findById(id).map(video -> {
-            video.setTitle(newVideo.getTitle());
-            video.setDescription(newVideo.getDescription());
-            video.setDuration(newVideo.getDuration());
-            video.setThumbnailUrl(newVideo.getThumbnailUrl());
-            video.setVideoUrl(newVideo.getVideoUrl());
-            return videoRepository.save(video);
-        }).orElseThrow(() -> new AppException(ErrorCode.VIDEO_NOT_EXISTED));
+    public VideoResponse updateVideo(String videoId, VideoUpdationRequest request) {
+        Video video = videoRepository.findById(videoId).orElseThrow(()-> new AppException(ErrorCode.VIDEO_NOT_EXISTED));
+        if(!video.getProfileId().equals(getCurrentUserId())){
+            throw new AppException(ErrorCode.UNAUTHORIZED);
+        }
+        video.setThumbnailUrl(request.getThumbnailUrl());
+        video.setDescription(request.getDescription());
+        video.setTitle(request.getTitle());
+        return videoMapper.toVideoResponse(videoRepository.save(video));
     }
 
     public void deleteVideo(String id) {
         Video video = videoRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.VIDEO_NOT_EXISTED));
+        if(!video.getProfileId().equals(getCurrentUserId())){
+            throw new AppException(ErrorCode.UNAUTHORIZED);
+        }
         videoRepository.delete(video);
     }
 
     public VideoResponse createVideo(MultipartFile videoFile, MultipartFile thumbnailFile, VideoCreationRequest request) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentProfileId = getCurrentUserId();
 
         Video video = Video.builder()
                 .title(request.getTitle())
@@ -85,7 +96,7 @@ public class VideoService {
         video.setVideoUrl(fileClient.uploadFile(videoFile));
         video.setThumbnailUrl(fileClient.uploadFile(thumbnailFile));
         video.setPublishedAt(LocalDateTime.now());
-        video.setUserId(authentication.getName());
+        video.setProfileId(currentProfileId);
         double durationInSeconds = 0.0;
         try {
             // convert MultipartFile -> File tạm
@@ -113,6 +124,9 @@ public class VideoService {
             grabber.stop();
             return durationInSeconds;
         }
+    }
+    private String getCurrentUserId(){
+        return SecurityContextHolder.getContext().getAuthentication().getName();
     }
